+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Remove `Converger.allConnectorsSynchronized()` as it was unused outside of tests and could return an incorrect
   result as not all dependencies were reactive.
+* Update the `org.realityforge.react4j` dependencies to version `0.126`.
 
 ### [v6.42](https://github.com/replicant4j/replicant/tree/v6.42) (2019-04-25)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.41...v6.42)